api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.18.4",
    "fastapi[standard]>=0.135.1",
    "jinja2>=3.1.6",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "redis>=7.2.1",
    "rich>=14.3.3",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]

[tool.scripts]

dev = "uvicorn main:app --reload --host 0.0.0.0 --port 8000"

db-current = "alembic current"
db-history = "alembic history"
db-heads = "alembic heads"

db-up = "alembic upgrade head"
db-down-1 = "alembic downgrade -1"

db-rev = "alembic revision -m"
db-auto = "alembic revision --autogenerate -m"

seed = "python -m scripts.seed"
